\name{JDU accessors}
\alias{anchorc}
\alias{anchorc<-}
\alias{localec}
\alias{localec<-}
\alias{anchorj}
\alias{anchorj<-}
\alias{localej}
\alias{localej<-}
\alias{jir}
\alias{jir<-}
\alias{jes}
\alias{jes<-}
\alias{jalt}
\alias{jalt<-}

\title{Accessors for ASpliJDU object}

\description{ Accessors for ASpliJDU object }
\usage{ 
  anchorc( x )
  anchorj( x )
  localec( x )
  localej( x )
  jir( x )
  jes( x )
  jalt( x )
}

\arguments{
  \item{ x }{ An ASpliJDU object }
}

\value{ Returns dataframes  }

\author{ Estefania Mancini, Andres Rabinovich, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }



