\name{splicingReport accessors}
\alias{binbased}
\alias{binbased<-}
\alias{localebased}
\alias{localebased<-}
\alias{anchorbased}
\alias{anchorbased<-}


\title{Accessors for ASpliSplicingReport object}

\description{ Accessors for ASpliSplicingReport object }
\usage{ 
  binbased( x )
  localebased( x )
  anchorbased( x )
}

\arguments{
  \item{ x }{ An ASpliSplicingReport object }
}

\value{ Returns dataframes }


\author{ Estefania Mancini, Andres Rabinovich, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }



