\name{ASpli-deprecated}
\alias{ASpli-deprecated}
\title{Deprecated functions in package \sQuote{ASpli}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{ASpli} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{loadBAM: \code{\link{gbCounts}}}
    \item{readCounts: \code{\link{gbCounts}}}
    \item{AsDiscover: \code{\link{jCounts}}}
      \item{DUreport: \code{\link{newFunc}}}
  
      }
  
   The following functions are deprecated and will be made defunct without replacement; 
  \itemize{
  
    \item{mergeBinDUAS}
    \item{mergeBinDUAS}
    \item{mergeBinDUAS}
    \item{plotGenomicRegions}
    

  }
  
}