\name{ASpli-AS-class}
\Rdversion{1.1}
\docType{class}
\alias{ASpli-AS-class}
\alias{altPSI,ASpli-AS-method}
\alias{esPSI,ASpli-AS-method}
\alias{irPIR,ASpli-AS-method}
\alias{joint,ASpli-AS-method}
\alias{junctionsPIR,ASpli-AS-method}
\alias{junctionsPSI,ASpli-AS-method}
\alias{writeAS,ASpli-AS-method}
\title{Class \code{"ASpli-AS"}}
\description{}
\section{Slots}{
\describe{
    \item{\code{irPIR}}{}
    \item{\code{altPSI}}{}
    \item{\code{esPSI}}{}
    \item{\code{junctionsPIR}}{}
    \item{\code{junctionsPSI}}{}
    \item{\code{join}}{}
    }
    }
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }