\name{ASpli-counts}
\docType{class}
\alias{ASpli-counts}
\description{ Contains results of read overlaps against all feature levels summarization }
\title{Class \code{"ASpli-counts"}}
\section{Slots}{
  \describe{
    \item{\code{gene.counts}}{}
    \item{\code{exon.intron.counts}}{}
    \item{\code{junction.counts}}{}
    \item{\code{e1i.counts}}{}
    \item{\code{ie2.counts}}{}
    \item{\code{gene.rd}}{}
    \item{\code{bin.rd}}{}
  }
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz}
