\name{ASpli-deprecated}
\alias{ASpli-deprecated}
\title{Deprecated functions in package \sQuote{ASpli}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{ASpli} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{
    \item{loadBAM: \code{\link{gbCounts}}}
    
    \item{readCounts: \code{\link{gbCounts}}}
    
    \item{AsDiscover: \code{\link{jCounts},
                            \link{splicingReport},    
                            \link{integrateSignals}}}
                            
    \item{DUreport: \code{\link{gbDUreport},
                          \link{jDUreport}}}

    \item{DUreportBinSplice: \code{\link{gbDUreport}}}
    \item{junctionDUreport: \code{\link{jDUreport}}}        
                        
    \item{mergeBinDUAS: \code{\link{splicingReport},
                              \link{integrateSignals}}}
                              
    \item{junctionsPSI: \code{\link{junctionsPJU}}}
    
    \item{plotGenomicRegions: \code{\link{exportSplicingReports}, \link{exportIntegratedSignals}}}   
        

}
    
}
 


 
