\name{ASpliFeatures-class}

\Rdversion{1.1}

\docType{class}

\alias{ASpliFeatures-class}
\alias{featuresb,ASpliFeatures-method}
\alias{featuresg,ASpliFeatures-method}
\alias{featuresj,ASpliFeatures-method}
\alias{readCounts,ASpliFeatures-method}
\alias{plotGenomicRegions,ASpliFeatures-method}
\alias{featuresb<-,ASpliFeatures,GRanges-method}
\alias{featuresg<-,ASpliFeatures,GRangesList-method}
\alias{featuresj<-,ASpliFeatures,GRanges-method}
\title{Class \code{"ASpliFeatures"}}

\description{Contains Genomic Ranges of different features extracted from a 
  TxDb}

\section{Slots}{
  \describe{
    \item{\code{genes}:}{}
    \item{\code{bins}:}{}
    \item{\code{junctions}:}{}
  }
}

\author{ Estefania Mancini, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz}
