\name{ASpliMergedReports-class}

\Rdversion{1.1}

\docType{class}

\alias{ASpliMergedReports}
\alias{ASpliMergedReports-class}
\alias{mergedReports,ASpliMergedReports-method}

\title{Class \code{"ASpliMergedReports"}}

\description{ Contains results of differential expression at junction level.
}

\section{Slots}{
  \describe{
    \item{\code{binbased}}{}
    \item{\code{localebased}}{}
    \item{\code{anchorbased}}{}
  }
}

\author{ Andres Rabinovich, Estefania Mancini, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }