\name{binGenome-methods}
\docType{methods}
\alias{binGenome-methods}
\alias{binGenome,TxDb-method}
\title{ Feature coordinates extraction }
\description{
Feature coordinates extraction from a Transcript Db Database
}
\section{Methods}{
\describe{
\item{\code{signature(genome = "TxDb")}}{
An object of class transcriptDb (TxDb) 
}
}}
\author{Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz }
\seealso{ \code{\link{featuresg}, \link{featuresb} , \link{featuresj} 
} 
}

