\name{Example data}

\alias{aspliASexample}
\alias{aspliBamsExample}
\alias{aspliCountsExample}
\alias{aspliDUexample1}
\alias{aspliDUexample2}
\alias{aspliExampleBamList}
\alias{aspliExampleGTF}
\alias{aspliFeaturesExample}
\alias{aspliJunctionDUexample}
\alias{aspliTargetsExample}


\title{
  Example Aspli objects
}
\description{
  ASpli includes functions to easily build ASpli objects, used in examples in 
  the vignette and man pages.
}

\usage{
  aspliASexample()
  aspliBamsExample()
  aspliCountsExample()
  aspliDUexample1()
  aspliDUexample2()
  aspliExampleBamList()
  aspliExampleGTF()
  aspliFeaturesExample()
  aspliJunctionDUexample()
  aspliTargetsExample()
}

\value{ An ASpli object with example data.}

\author{ Estefania Mancini, Andres Rabinovich, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }

\examples{
  #as <- aspliASexample()
  #bams <- aspliBamsExample()
  #counts <- aspliCountsExample()
  #du1 <- aspliDUexample1()
  #du2 <- aspliDUexample2()
  #bamfiles <- aspliExampleBamList()
  #gtffile <- aspliExampleGTF()
  #features <- aspliFeaturesExample()
  #jdu <- aspliJunctionDUexample()
  #targets <-aspliTargetsExample()
}


