\name{rds}
\alias{rds}
\title{
Divides read counts by gene and bin length
}
\description{
Divides read counts by gene and bin length
}
\usage{
rds(counts, targets)
}
\arguments{
  \item{counts}{
An ASpliCounts object
}
  \item{targets}{
Target dataframe
}
}
\value{
Read densities of genes and bins
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }
