\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,ASpliAS-method}
\alias{show,ASpliCounts-method}
\alias{show,ASpliDU-method}
\alias{show,ASpliFeatures-method}
\title{ Display a summary of data contained in ASpliObjects }
\description{ Display a summary of data contained in ASpliObjects}
\details{ Display a summary of data contained in ASpliObjects }
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }
