\name{MergedReports accessors}
\alias{binbased}
\alias{binbased<-}
\alias{localebased}
\alias{localebased<-}
\alias{anchorbased}
\alias{anchorbased<-}


\title{Accessors for ASpliMergedReport object}

\description{ Accessors for ASpliMergedReport object }
\usage{ 
  binbased( x )
  localebased( x )
  anchorbased( x )
}

\arguments{
  \item{ x }{ An ASpliMergedReport object }
}

\value{ Returns dataframes }

\author{ Estefania Mancini, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz }



