\name{ASpli-DU-class}
\Rdversion{1.1}
\docType{class}
\alias{ASpli-DU-class}
\alias{binsDU,ASpli-DU-method}
\alias{genesDE,ASpli-DU-method}
\alias{junctionsDU,ASpli-DU-method}
\alias{writeDU,ASpli-DU-method}
\title{Class \code{"ASpli-DU"}}
\description{}
\section{Slots}{
  \describe{
    \item{\code{genes}}{}
    \item{\code{bins}}{}
    \item{\code{junctions}}{}
  }
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }