\name{ASpli-features-class}
\Rdversion{1.1}
\docType{class}
\alias{ASpli-features-class}
\alias{featuresb,ASpli-features-method}
\alias{featuresg,ASpli-features-method}
\alias{featuresj,ASpli-features-method}
\alias{readCounts,ASpli-features-method}
\title{Class \code{"ASpli-features"}}
\description{}
\section{Slots}{
\describe{
    \item{\code{genes}:}{}
    \item{\code{bins}:}{}
    \item{\code{junctions}:}{}
  }
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }
