  \name{ASpliCounts}

\docType{class}

\alias{ASpliCounts}

\description{ Contains results of read overlaps against all feature levels 
  summarization }

\title{ Class \code{"ASpliCounts"} }

\section{Slots}{
  \describe{
    \item{\code{gene.counts}}{}
    \item{\code{exon.intron.counts}}{}
    \item{\code{junction.counts}}{}
    \item{\code{e1i.counts}}{}
    \item{\code{ie2.counts}}{}
    \item{\code{gene.rd}}{}
    \item{\code{bin.rd}}{}
    \item{\code{condition.order}}{}
  }
}

\author{ Estefania Mancini, Andres Rabinovich, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }
