\name{ASpliDU-class}
\Rdversion{1.1}
\docType{class}
\alias{ASpliDU-class}
\alias{binsDU,ASpliDU-method}
\alias{genesDE,ASpliDU-method}
\alias{junctionsDU,ASpliDU-method}
\alias{writeDU,ASpliDU-method}
\title{Class \code{"ASpliDU"}}
\description{Contains results of differential expression at gene level and differential usage at bin and junction level estimation using DEreport method.
}
\section{Slots}{
  \describe{
    \item{\code{genes}}{}
    \item{\code{bins}}{}
    \item{\code{junctions}}{}
  }
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }