\name{ASpliFeatures-class}
\Rdversion{1.1}
\docType{class}
\alias{ASpliFeatures-class}
\alias{featuresb,ASpliFeatures-method}
\alias{featuresg,ASpliFeatures-method}
\alias{featuresj,ASpliFeatures-method}
\alias{readCounts,ASpliFeatures-method}
\title{Class \code{"ASpliFeatures"}}
\description{Contains Genomic Ranges of different features extracted from a     TxDb}
\section{Slots}{
\describe{
    \item{\code{genes}:}{}
    \item{\code{bins}:}{}
    \item{\code{junctions}:}{}
  }
}
\author{ Estefania Mancini, Marcelo Yanovsky, Ariel Chernomoretz  }
