\name{ASpliIntegratedSignals-class}

\Rdversion{1.1}

\docType{class}

\alias{ASpliIntegratedSignals}
\alias{ASpliIntegratedSignals-class}
\alias{anchorbased,ASpliIntegratedSignals-method}
\alias{binbased,ASpliIntegratedSignals-method}
\alias{localebased,ASpliIntegratedSignals-method}
\alias{anchorbased<-,ASpliIntegratedSignals,data.frame-method}
\alias{localebased<-,ASpliIntegratedSignals,data.frame-method}
\alias{binbased<-,ASpliIntegratedSignals,data.frame-method}
\alias{splicingReport,ASpliIntegratedSignals-method}
\alias{show,ASpliIntegratedSignals-method}
\alias{writeIntegratedSignals,ASpliIntegratedSignals-method}


\title{Class \code{"ASpliIntegratedSignals"}}

\description{ Contains results of differential expression at junction level.
}

\section{Slots}{
  \describe{
    \item{\code{binbased}}{}
    \item{\code{localebased}}{}
    \item{\code{anchorbased}}{}
  }
}

\author{ Andres Rabinovich, Estefania Mancini, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }