\name{ASpliJDU-class}

\Rdversion{1.1}

\docType{class}

\alias{ASpliJDU}
\alias{ASpliJDU-class}
\alias{junctionsDUreport,ASpliJDU-method}

\title{Class \code{"ASpliJDU"}}

\description{ Contains results of differential expression at junction level.
}

\section{Slots}{
  \describe{
    \item{\code{localec}}{}
    \item{\code{localej}}{}
    \item{\code{anchorc}}{}
    \item{\code{anchorj}}{}
    \item{\code{jir}}{}
    \item{\code{jes}}{}
    \item{\code{jalt}}{}
    \item{\code{contrast}}{}    
  }
}

\author{ Andres Rabinovich, Estefania Mancini, Javier Iserte, Marcelo Yanovsky, Ariel Chernomoretz  }